renamed `base_weartime_detector` to `base`
